pastebin rework
